convert to fp16 in _fwd_kernel_alibi

+++ vllm/model_executor/layers/triton_kernel/prefix_prefill.py
@@ -481,7 +481,7 @@ def _fwd_kernel_alibi(
                  (start_n + offs_n[:, None]) % block_size * stride_v_cache_bl)
         k = tl.load(K_cache + off_k,
                     mask=(start_n + offs_n[None, :]) < cur_batch_ctx_len,
-                    other=0.0)
+                    other=0.0).to(q.dtype)
 
         qk = tl.zeros([BLOCK_M, BLOCK_N], dtype=tl.float32)
         qk += tl.dot(q, k)
@@ -516,7 +516,7 @@ def _fwd_kernel_alibi(
         # update acc
         v = tl.load(V_cache + off_v,
                     mask=(start_n + offs_n[:, None]) < cur_batch_ctx_len,
-                    other=0.0)
+                    other=0.0).to(q.dtype)
 
         p = p.to(v.dtype)
         acc += tl.dot(p, v, allow_tf32=False)
